*Minor changes to satisfy cppcheck and Splint warnings. *Altered RS2 export to prevent duplicate single notes from exporting in the event of overlapping arpeggio phrases. *Added logic to remove the possibility of overlapping arpeggio phrases. *Removed unused functions eof_song_tick_to_msec() and eof_song_msec_to_tick(). *Changed eof_get_beat_length() to use the floating point beat positions, since the calling function was using floating point math anyway.  This should improve the accuracy of the auto-adjust logic. *Changed the handling of time signatures so that denominator (beat unit) alters the length of a beat as it does in music theory. *Removed the double float typecast from numbers that can just be written with a decimal point, since that's the type they are considered to be in C.

+++ src/editor.c
@@ -5009,7 +5009,7 @@ int eof_get_tempo_text(int beat, char * buffer)
 		return 0;
 	}
 
-	current_bpm = (double)60000000.0 / (double)eof_song->beat[beat]->ppqn;
+	current_bpm = 60000000.0 / (double)eof_song->beat[beat]->ppqn;
 	(void) uszprintf(buffer, 16, "%03.2f ", current_bpm);
 
 	return 1;
@@ -5695,12 +5695,12 @@ void eof_render_editor_window_common(EOF_WINDOW *window)
 
 			if(i == eof_selected_beat)
 			{	//Draw selected beat's tempo
-				current_bpm = (double)60000000.0 / (double)eof_song->beat[i]->ppqn;
+				current_bpm = 60000000.0 / (double)eof_song->beat[i]->ppqn;
 				textprintf_ex(window->screen, eof_mono_font, xcoord - 28, EOF_EDITOR_RENDER_OFFSET + 7 - (i % 2 == 0 ? 0 : 10), i == eof_hover_beat ? bhcol : i == eof_selected_beat ? bscol : bcol, -1, "<%03.2f>", current_bpm);
 			}
 			else if(eof_song->beat[i]->contains_tempo_change)
 			{	//Draw tempo if this beat has a tempo change
-				current_bpm = (double)60000000.0 / (double)eof_song->beat[i]->ppqn;
+				current_bpm = 60000000.0 / (double)eof_song->beat[i]->ppqn;
 				textprintf_ex(window->screen, eof_mono_font, xcoord - 20, EOF_EDITOR_RENDER_OFFSET + 7 - (i % 2 == 0 ? 0 : 10), i == eof_hover_beat ? bhcol : i == eof_selected_beat ? bscol : bcol, -1, "%03.2f", current_bpm);
 			}
 			else
